refactor withRetry func to utils
